+++ firmware/CHANGELOG.md
@@ -33,6 +33,7 @@ or
 ### Added
  - Staged injection: a second set of fuel injectors that inject part of the fuel load when you run out of primary injector
  - VE/Ign/Boost blend tables can select a Y axis override for even more flexibility
+ - Added Lua function `getAuxDigital` for parity with rusEFI #351
 
 ## December 2023 Release